fix(dictfile & co): no more variants pollution
We are using a mechanism to support variants with different prefixes for Kobo (DictHTML output format).
The same mechanism was applied to DictFile and its sub-formats, resulting in dictionaries with lot
of "pollution" from variants being included with duplicate definitions.

I think the workaround done for Kobo should not apply to to DictFile, and sub-formats.
Moreover, the dictFile format specification allowes variants with different prefixes:
cf https://pgaskin.net/dictutil/dictgen/#dictfile-format