fix(dictfile & co): no more variants pollution (#2487)
We are using a mechanism to support variants with different prefixes for Kobo (DictHTML output format). The same mechanism was applied to DictFile and its sub-formats, resulting in dictionaries with lot of "pollution" from variants being included with duplicate definitions.

I think the workaround done for Kobo should not apply to DictFile, and sub-formats.