fix: --render is now really running in parallel
Multiprocessing was not really behaved as we intended, and all words were duplicated to all pools.
Sometimes, no duplication was done, and other times a word would be found in 2-3 pools.

The new code does a word list split by the number of workers, and distribute the workload
manually. Performances are now crazy (divided by 3-5)!

Related to #2528.