feat/fix: rework --parse and --render
`--parse` now save all words in the database, without filtering.
This allowes to parse (big) XML files only once (required for multilingual velocity),
and to handle properly redirections: no more foreign word redirection polluting the dictionary.
Also, handling of redirections is faster, since it is one SQL select + a dict iteration.
As a result, `data_wikicode-YYYYMMDD.json` files are gone; and the (big) XML file is deleted on success.

`--render` no more have to deal with words redirection, and loading words from the database is way faster.

Here is the file tree for a dictionary after `--download`, `--parse`, and `--render`:

    data/LOCALE
    ├── LOCALE
    │   └── data-20251001.json
    ├── pages-20251001.sqlite
    └── pages-20251001.xml.bz2

    2 directories, 3 files

Where before it was like:

    data/LOCALE
    ├── LOCALE
    │   ├── data-20251001.json
    │   ├── data_wikicode-20251001.json
    │   └── modules-20251001.sqlite
    ├── pages-20251001.xml
    └── pages-20251001.xml.bz2

    2 directories, 5 files